security update

+++ routes/index.js
@@ -26,6 +26,7 @@ router.get('/cbuilder', ensureAuthenticated, async(req, res) => {
 })
 
 router.get('/game', ensureAuthenticated, async(req, res) => {
+    if (req.user.admin == false) return res.redirect("/403")
     if (req.useragent.isMobile == true) return res.render('mobile')
     res.render('index', {
         User: req.user
@@ -53,6 +54,7 @@ router.get('/player', ensureAuthenticated, async(req, res) => {
 })
 
 router.get('/terminal', ensureAuthenticated, async(req, res) => {
+    if (req.user.terminal == false) return res.redirect("/403")
     if (req.useragent.isMobile == true) return res.render('mobile')
     Users.find({}, function(err, users) {
         Classes.find({}, function(err, classes) {
@@ -69,11 +71,13 @@ router.get('/terminal', ensureAuthenticated, async(req, res) => {
 })
 
 router.get('/terminal/deleteuser/:id', ensureAuthenticated, async(req, res) => {
+    if (req.user.terminal == false) return res.redirect("/403")
     await Users.findOneAndDelete({_id: req.params.id})
     res.redirect("/terminal")
 })
 
 router.get('/terminal/resetuser/:id', ensureAuthenticated, async(req, res) => {
+    if (req.user.terminal == false) return res.redirect("/403")
     await Users.findOneAndUpdate(
         { _id: req.params.id },
         { class: "none", weapon: "none" }
@@ -82,17 +86,20 @@ router.get('/terminal/resetuser/:id', ensureAuthenticated, async(req, res) => {
 })
 
 router.get('/terminal/deleteclass/:id', ensureAuthenticated, async(req, res) => {
+    if (req.user.terminal == false) return res.redirect("/403")
     await Classes.findOneAndDelete({_id: req.params.id})
     res.redirect("/terminal")
 })
 
 router.get('/terminal/deleteweapon/:id', ensureAuthenticated, async(req, res) => {
+    if (req.user.terminal == false) return res.redirect("/403")
     await Weapons.findOneAndDelete({_id: req.params.id})
     res.redirect("/terminal")
 })
 
 
 router.get('/devices', ensureAuthenticated, async(req, res) => {
+    if (req.user.terminal == false) return res.redirect("/403")
     if (req.useragent.isMobile == true) return res.render('mobile')
     Versions.findOne({program: "weapon"}, function(err, weapon) {
         Versions.findOne({program: "vest"}, function(err, vest) {
